Extract the formatter as a behaviour to allow for multiple Formatter types
